refactor: : Socket controller for new notification
- move CRUD logic related to new notification creation and data validation to new controller factory function `createNewNotification`, from `handleNotifications.js` to `notificationControllers.js`.
- add additional validation and support for factory models and error handling in `createNewNotification` in `examControllers.js`.
- change handler function from `createNewNotification ` to `handleCreateNewNotification ` in `handleNotifications.js`.
